Job c параметрами #3
Реализовал функционал.

+++ db/scripts/rabbit_schemas.sql
@@ -1,4 +1,4 @@
 CREATE TABLE IF NOT EXISTS rabbit(
 id serial PRIMARY KEY,
-created_date int
+created_date bigint
 );
